backend/requirements.txt
# SentimentDNA API Backend Requirements
# =====================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# WebSocket Support
websockets==12.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# Monitoring & Metrics
prometheus-client==0.19.0
structlog==24.1.0

# HTTP Client (for external APIs)
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
numpy==1.26.3
pandas==2.1.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Development
black==24.1.0
ruff==0.1.13
mypy==1.8.0

